% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_prob.R
\name{cal_prob}
\alias{cal_prob}
\title{Calculate probabilities of a vote matrix}
\usage{
cal_prob(vote, post_samples)
}
\arguments{
\item{post_samples}{A list of posterior samples of parameters obtained from MCMC.}

\item{vote_m}{A vote matrix in which rows represent members and columns represent issues.
The entries can only be 0 (indicating ”No”), 1 (indicating ”Yes”), or NA (indicating missing data).}
}
\value{
A matrix of probabilities corresponding to each vote
}
\description{
This function is used to get the probabilities of each vote result for the given members and issues
}
\examples{
prob_mat = cal_prob(vote = house_votes_m, post_samples = post_samples)
}
