% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_waic_static.R
\name{cal_row_likelihood}
\alias{cal_row_likelihood}
\title{calculate the log-likelihood of each row (input vectors)}
\usage{
cal_row_likelihood(vote, beta, alpha1, alpha2, delta1, delta2)
}
\description{
calculate the log-likelihood of each row (input vectors)
}
