% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_waic.R
\name{cal_waic}
\alias{cal_waic}
\title{Calculate WAIC}
\usage{
cal_waic(vote, post_samples)
}
\arguments{
\item{post_samples}{A list of posterior samples of parameters obtained from MCMC.}

\item{vote_m}{A vote matrix in which rows represent members and columns represent issues.
The entries can only be 0 (indicating ”No”), 1 (indicating ”Yes”), or NA (indicating missing data).}
}
\value{
The WAIC value of this probit unfoldinng model
}
\description{
This function is used to get the WAIC value of the probit unfolding model
}
\examples{
waic_val = cal_waic(vote = house_votes_m, post_samples = post_samples)
}
