% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_waic_dynamic.R
\name{cal_waic_dynamic}
\alias{cal_waic_dynamic}
\title{Calculate WAIC of the probit unfolding static model}
\usage{
cal_waic_dynamic(vote_m, years_v, chain_run)
}
\arguments{
\item{vote_m}{A vote matrix in which rows represent members and columns represent issues.
The entries can only be 0 (indicating ”No”), 1 (indicating ”Yes”), or NA (indicating missing data).}

\item{years_v}{A time vector}

\item{chain_run}{The posterior samples obtained from MCMC}
}
\description{
This function is used to get the WAIC value of the static probit unfolding model
}
\examples{
waic_dy = cal_waic_dynamic(mqVotes, mqTime, post_samples_dy)
}
