% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_waic_static.R, R/tune_hyper.R
\name{get_prob_mat}
\alias{get_prob_mat}
\title{calculate probability matrix of a given set of parameters}
\usage{
get_prob_mat(vote, beta, alpha1, alpha2, delta1, delta2)

get_prob_mat(vote, beta, alpha1, alpha2, delta1, delta2)
}
\description{
calculate probability matrix of a given set of parameters

calculate probability matrix of a given set of parameters
}
