% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_dynamic.R
\name{init_data_gp_rcpp}
\alias{init_data_gp_rcpp}
\title{initialize member and position parameters and get starting points}
\usage{
init_data_gp_rcpp(
  vote_m,
  years_v,
  leg_pos_init,
  alpha_pos_init,
  delta_pos_init,
  rho_init,
  y_star_m_1_init,
  y_star_m_2_init,
  y_star_m_3_init,
  total_iter,
  pos_ind_list,
  neg_ind_list,
  pos_ind_years_list,
  neg_ind_years_list
)
}
\description{
initialize member and position parameters and get starting points
}
