% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_three_utility_probit.R
\name{init_data_rcpp}
\alias{init_data_rcpp}
\title{initialize member and position parameters and get starting points}
\usage{
init_data_rcpp(
  vote_m,
  leg_pos_init,
  alpha_pos_init,
  delta_pos_init,
  y_star_m_1_init,
  y_star_m_2_init,
  y_star_m_3_init,
  total_iter
)
}
\description{
initialize member and position parameters and get starting points
}
