% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_char_curve.R
\name{item_char_curve}
\alias{item_char_curve}
\title{Produces the item characteristic curves}
\usage{
item_char_curve(rollnumber, post_samples)
}
\arguments{
\item{rollnumber}{The roll number of the issue to be reviewed.}

\item{post_samples}{A list of posterior samples of parameters obtained from MCMC.}
}
\value{
An item characteristic curve of the input issue
}
\description{
This function is used to generate the item characteristic curve for a given issue
}
\examples{
item_result = item_char_curve(rollnumber = 100, post_samples = post_samples)
}
