% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_dynamic.R
\name{sample_probit_dynamic}
\alias{sample_probit_dynamic}
\title{generate posterior samples from the dynamic probit unfolding model}
\usage{
sample_probit_dynamic(
  vote_m,
  years_v,
  hyperparams,
  iter_config,
  pos_ind_list = NULL,
  neg_ind_list = NULL,
  pos_ind_years_list = NULL,
  neg_ind_years_list = NULL
)
}
\arguments{
\item{vote_m}{A vote matrix in which rows represent members and columns represent issues.
The entries can only be 0 (indicating ”No”), 1 (indicating ”Yes”), or NA (indicating missing data).}

\item{years_v}{A time vector}

\item{hyperparams}{A list of hyperparameter values. beta_mean is a value for the prior mean of beta.
beta_var is a value for the variance of beta. alpha_mean is a vector of 2 components representing the prior
means of alpha1 and alpha2. alpha_scale is a value for the scale parameters of alpha1 and alpha2.
delta_mean is a vector of 2 components representing the prior means of delta1 and delta2.
delta_scale is a value for the scale parameters of delta1 and delta2. nu is a fixed value making phi follow
Gamma(nu/2,nu/2).}

\item{iter_config}{A list corresponding to iteration configurations, num_iter is the total number of iterations.
start_iter is the iteration number to start retaining data after the burn-in period. keep_iter is the interval
at which iterations are kept for posterior samples. flip_rate is the probability of fliping signs directly in the M-H step.}

\item{pos_ind_list}{Index of members set to be positive}

\item{neg_ind_list}{Index of members set to be negative}

\item{pos_ind_years_list}{Index of corresponding years set to be positive}

\item{neg_ind_years_list}{Index of corresponding years set to be negative}

\item{leg_pos_init}{A vector of initial values for all the positions of legislators. The
length of this vector should be equal to the number of iterations times the number of members}

\item{alpha_pos_init}{A vector of initial values for all the alpha parameters. This vector
should have a length of 2 times the number of iterations times the number of issues,
as each issue has two alpha parameters.}

\item{delta_pos_init}{A vector of initial values for all the delta parameters. The length
should be 2 times the number of iterations times the number of issues}

\item{y_star_m_1_init}{Matrices of initial values for the three auxiliary variables, with
dimensions equal to the number of members by the number of issues. Same for the other two parameters.}

\item{start_val}{A vector of all initial values for each parameter.}
}
\value{
A list mainly containing:
- `beta`: A matrix of parameter draws for beta.
- `alpha1`: A matrix of parameter draws of alpha1.
- `alpha2`: A matrix of parameter draws of alpha2.
- `delta1`: A matrix of parameter draws of delta1.
- `delta2`: A matrix of parameter draws of delta2.
- `rho`: A vector of parameter draws of rho.
}
\description{
This function is used to get samples of all the parameters from their posterior distributions
based on the dynamic probit unfolding model
}
\examples{
hyperparams = list(beta_mean = 0, beta_var = 1, alpha_mean = c(0, 0), alpha_scale = 5,
                    delta_mean = c(-2, 10), delta_scale = sqrt(10),
                    rho_mean = 0.9, rho_sigma = 0.04, rho_sd = 0.1, nu = 1)
iter_config = list(num_iter = 100, start_iter = 0, keep_iter = 1, flip_rate = 0.1)
pos_inds = c(39, 5)
neg_inds = c(12, 29)
pos_year_inds = list(1:31, 1)
neg_year_inds = list(1:29, 1:24)
post_samples = sample_probit_dynamic(mqVotes, mqTime, hyperparams, iter_config,
                pos_ind_list = pos_inds, pos_ind_years_list = pos_year_inds,
                neg_ind_list = neg_inds, neg_ind_years_list = neg_year_inds)
}
