% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_static.R
\name{sample_probit_static}
\alias{sample_probit_static}
\title{Generate posterior samples from the static probit unfolding model}
\usage{
sample_probit_static(vote_m, hyperparams, iter_config, pos_ind = 0)
}
\arguments{
\item{vote_m}{A vote matrix in which rows represent members and columns represent issues.
The entries can only be 0 (indicating ”No”), 1 (indicating ”Yes”), or NA (indicating missing data).}

\item{hyperparams}{A list of hyperparameter values. beta_mean is a value for the prior mean of beta.
beta_var is a value for the variance of beta. alpha_mean is a vector of 2 components representing the prior
means of alpha1 and alpha2. alpha_scale is a value for the scale parameters of alpha1 and alpha2.
delta_mean is a vector of 2 components representing the prior means of delta1 and delta2.
delta_scale is a value for the scale parameters of delta1 and delta2. nu is a fixed value making phi follow
Gamma(nu/2,nu/2).}

\item{iter_config}{A list corresponding to iteration configurations, num_iter is the total number of iterations.
start_iter is the iteration number to start retaining data after the burn-in period. keep_iter is the interval
at which iterations are kept for posterior samples. flip_rate is the probability of fliping signs directly in the M-H step.}

\item{pos_ind}{The index of the reference member whose position is kept positive.}

\item{leg_pos_init}{A vector of initial values for all the positions of legislators. The
length of this vector should be equal to the number of iterations times the number of members}

\item{alpha_pos_init}{A vector of initial values for all the alpha parameters. This vector
should have a length of 2 times the number of iterations times the number of issues,
as each issue has two alpha parameters.}

\item{delta_pos_init}{A vector of initial values for all the delta parameters. The length
should be 2 times the number of iterations times the number of issues}

\item{y_star_m_1_init}{Matrices of initial values for the three auxiliary variables, with
dimensions equal to the number of members by the number of issues. Same for the other two parameters.}

\item{start_val}{A vector of all initial values for each parameter.}

\item{sample_beta}{A Boolean value indicating whether to sample beta in the model.}
}
\value{
A list mainly containing:
- `beta`: A matrix of parameter draws for beta.
- `alpha1`: A matrix of parameter draws of alpha1.
- `alpha2`: A matrix of parameter draws of alpha2.
- `delta1`: A matrix of parameter draws of delta1.
- `delta2`: A matrix of parameter draws of delta2.
}
\description{
This function is used to get samples of all the parameters from their posterior distributions
based on the static probit unfolding model
}
\examples{
hyperparams = list(beta_mean = 0, beta_var = 1, alpha_mean = c(0, 0),
                   alpha_scale = 5, delta_mean = c(-2, 10), delta_scale = sqrt(10), nu = 10)
iter_config = list(num_iter = 100, start_iter = 0, keep_iter = 1, flip_rate = 0.1)

post_samples = sample_probit_static(house_votes_m, hyperparams, iter_config, pos_ind = 82)
}
