% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_three_utility_probit.R
\name{sample_three_utility_probit}
\alias{sample_three_utility_probit}
\title{generate posterior samples}
\usage{
sample_three_utility_probit(
  vote_m,
  hyperparams,
  num_iter = 20000,
  start_iter = 10000,
  keep_iter = 1,
  leg_pos_init = NULL,
  alpha_pos_init = NULL,
  delta_pos_init = NULL,
  y_star_m_1_init = NULL,
  y_star_m_2_init = NULL,
  y_star_m_3_init = NULL,
  pos_ind = 0,
  neg_ind = 0,
  start_val = NULL,
  sample_beta = T
)
}
\arguments{
\item{vote_m}{A vote matrix in which rows represent members and columns represent issues.
The entries can only be 0 (indicating ”No”), 1 (indicating ”Yes”), or NA (indicating missing data).}

\item{hyperparams}{A list of hyperparameter values. beta_mean is a value for the prior mean of beta.
beta_var is a value for the variance of beta. alpha_mean is a vector of 2 components representing the prior
means of alpha1 and alpha2. alpha_var is a value for the variance of alpha1 and alpha2.
delta_mean is a vector of 2 components representing the prior means of delta1 and delta2.
delta_var is a value for the variance of delta1 and delta2.}

\item{num_iter}{The total number of iterations}

\item{start_iter}{The iteration number to start retaining data after the burn-in period}

\item{keep_iter}{The interval at which iterations are kept for posterior samples.}

\item{leg_pos_init}{A vector of initial values for all the positions of legislators. The
length of this vector should be equal to the number of iterations times the number of members}

\item{alpha_pos_init}{A vector of initial values for all the alpha parameters. This vector
should have a length of 2 times the number of iterations times the number of issues,
as each issue has two alpha parameters.}

\item{delta_pos_init}{A vector of initial values for all the delta parameters. The length
should be 2 times the number of iterations times the number of issues}

\item{y_star_m_1_init}{Matrices of initial values for the three auxiliary variables, with
dimensions equal to the number of members by the number of issues. Same for the other two parameters.}

\item{pos_ind}{The index of the reference member whose position is kept positive.}

\item{neg_ind}{The index of the reference member whose position is kept negative.}

\item{start_val}{A vector of all initial values for each parameter.}

\item{sample_beta}{A Boolean value indicating whether to sample beta in the model.}
}
\value{
A list mainly containing:
- `beta`: A matrix of parameter draws for beta.
- `alpha1`: A matrix of parameter draws of alpha1.
- `alpha2`: A matrix of parameter draws of alpha2.
- `delta1`: A matrix of parameter draws of delta1.
- `delta2`: A matrix of parameter draws of delta2.
}
\description{
This function is used to get samples of all the parameters from their posterior distributions
}
\examples{
hyperparams = list(leg_mean = 0, leg_s = 1, alpha_mean = c(0, 0), alpha_cov = 25,
                   delta_mean = c(-2, 10), delta_cov = 10)
post_samples = sample_three_utility_probit(house_votes_m, hyperparams,
          num_iter = 400, start_iter = 200,
          keep_iter = 1, leg_pos_init = NULL, alpha_pos_init = NULL,
          delta_pos_init = NULL, y_star_m_1_init = NULL, y_star_m_2_init = NULL,
          y_star_m_3_init = NULL, pos_ind = 82, neg_ind = 0, start_val = NULL,
          sample_beta = T)
}
