% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_hyper.R
\name{tune_hyper}
\alias{tune_hyper}
\title{generate posterior samples}
\usage{
tune_hyper(hyperparams = hyperparams, n_member, n_issue)
}
\arguments{
\item{hyperparams}{A list of hyperparameter values. beta_mean is a value for the prior mean of beta.
beta_var is a value for the variance of beta. alpha_mean is a vector of 2 components representing the prior
means of alpha1 and alpha2. alpha_var is a value for the variance of alpha1 and alpha2.
delta_mean is a vector of 2 components representing the prior means of delta1 and delta2.
delta_var is a value for the variance of delta1 and delta2.}

\item{n_member}{A value representing the number of members to be simulated}

\item{n_issue}{A value representing the number of issuess to be simulated}
}
\value{
A list containing:
- `samples`: A list containing all sampled parameters from given distributions.
- `probability`: A matrix of probabilities of voting "Yes"
- `histogram`: A histogram of the density of probabilities.
}
\description{
This function is used to get samples of all the parameters from their posterior distributions
}
\examples{
hyperparams = list(beta_mean = 0, beta_var = 1, alpha_mean = c(0, 0),
                   alpha_scale = 5, delta_mean = c(-2, 10),
                   delta_scale = sqrt(10), nu = 10)
tune_results = tune_hyper(hyperparams, n_member = 100, n_issue = 100)
}
